Added columns to games
Added api_id column to house the api_id, I will use this to validate
the uniqueness of a game with the scope of platform because a game with
multiple platforms, has the same api_id.
Added totalAdds column to be an accumulator corresponding to the
number of users that have added the game to their wanted_games table.
Put new columns in the serializer
Put validation logic in the game model