Unsuccessful post to games now reroutes to post to wanted_games
Made a new route (find) that is a GET request taking the api_id as a
parameter.
Coded the new route in the controller.
Error on post to game now reroutes to the find (show with api_id params)
to find the game within the games table so that the id can be used as
params for a post to wanted_games by the front end.