Changed a route and added exception to open_read_controller
The route for 'find' and 'show' both were '/games/:api_id' and
'/games/:id'. The API cannot differentiate between routes so it was
causing conflicts, especially since one was protected while the other
wasn't.
To fix this I changed the route for 'find' to 'game/:api_id'.
I also added the 'find' action to the list of actions that do not require
authentication in open_read_controller.rb.